6lesson/app/requirements.txt
SQLAlchemy==2.0.31
fastapi==0.116.1
uvicorn==0.30.1
bcrypt==4.0.1
dataclasses==0.6
sentencepiece==0.2.0
pydantic==2.7.1
pydantic-settings==2.1.0
asyncpg==0.29.0
torch==2.5.1
transformers==4.40.1
email-validator==2.2.0
aiogram==3.5.0
aiohttp==3.9.5
passlib[bcrypt]==1.7.4
sacremoses==0.1.1
python-jose==3.3.0
python-multipart==0.0.9
pika==1.3.2
jinja2==3.1.6
aiofiles==23.2.1
httpx==0.28.1
